refactor: move deleteRelayMessage function to a new module for better organization and update imports accordingly
