--- core/mpc/deleteRelayMessage.ts
@@ -1,29 +1,0 @@
-type DeleteRelayMessageInput = {
-  serverUrl: string
-  localPartyId: string
-  sessionId: string
-  messageHash: string
-  messageId?: string
-}
-export const deleteRelayMessage = async ({
-  serverUrl,
-  localPartyId,
-  sessionId,
-  messageHash,
-  messageId,
-}: DeleteRelayMessageInput): Promise<void> => {
-  const headers: Record<string, string> = {
-    'Content-Type': 'application/json',
-  }
-  if (messageId) {
-    headers['message_id'] = messageId
-  }
-
-  await fetch(
-    `${serverUrl}/message/${sessionId}/${localPartyId}/${messageHash}`,
-    {
-      method: 'DELETE',
-      headers,
-    }
-  )
-}
